Store all products in local hash
Add setAllProducts func to cartParse.js, which takes the array of
project objects that the API returns, and stores each object in a hash
called store.allProducts. Each hash entry has a key equal to the
product's ID, and a value equal to the entire product obj. This will
allow us to easily look up product info.

Ian